displaying txs fees better
