+++ src/components/explorer/tx/show-transaction.vue
@@ -132,7 +132,6 @@
                         <span class="col-5 col-sm-3 text-truncate">Fee Conversion Rate</span>
                         <span class="col-7 col-sm-9 text-truncate">{{ computeFeeRate(payload) }}</span>
                     </div>
-
                 </template>
 
                 <div class="row pt-2 pb-2 bg-light">
@@ -171,6 +170,21 @@
             </div>
 
         </template>
+
+        <div v-for="(fee,index) in fees"
+             :key="`fee_${index}`">
+
+            <div class="row pt-2 pb-2">
+                <span class="col-5 col-sm-3 fw-medium text-truncate">Fee{{ (fees.length > 1) ? index : ''}}</span>
+                <span class="col-7 col-sm-9 text-truncate">
+                    <amount :asset="fee.asset" :value="fee.amount" value-class="text-900" :sign="false" />
+                    <template v-if="fee.amountNative">
+                        => <amount :value="fee.amountNative" value-class="text-900" :sign="true" />
+                    </template>
+                </span>
+            </div>
+        </div>
+
     </div>
 </template>
 
@@ -179,10 +193,11 @@ import ShowTransactionData from "./show-transaction-data"
 import ShowTransactionDataExtra from "./show-transaction-data-extra"
 import StringHelper from "src/utils/string-helper";
 import Decimal from "decimal.js"
+import Amount from "src/components/wallet/amount"
 
 export default {
 
-    components: {ShowTransactionData, ShowTransactionDataExtra},
+    components: {ShowTransactionData, ShowTransactionDataExtra, Amount},
 
     props: {
         tx: {default: null},
@@ -193,6 +208,27 @@ export default {
     computed:{
 
         PandoraPay: () => PandoraPay,
+
+        fees(){
+            if (!this.tx) return []
+
+            if (this.tx.version.eq( PandoraPay.enums.transactions.TransactionVersion.TX_SIMPLE) ){
+                return [ {amount: this.tx.fee, asset: PandoraPay.config.coins.NATIVE_ASSET_FULL_STRING_HEX, feeRate: 1 } ]
+            }else if (this.tx.version.eq( PandoraPay.enums.transactions.TransactionVersion.TX_ZETHER) ) {
+
+                const out = []
+                for (const payload of this.tx.payloads)
+                    out.push({
+                        amount: payload.statement.fee,
+                        amountNative: payload.asset !== PandoraPay.config.coins.NATIVE_ASSET_FULL_STRING_HEX ? payload.statement.fee.mul(payload.feeRate).div(new Decimal(10).pow(payload.feeLeadingZeros)) : null,
+                        asset: payload.asset,
+                        feeRate: payload.feeRate
+                    })
+
+                return out
+            }
+
+        }
 
     },
