+++ src/components/send/txs/confirm-broadcasting-tx.vue
@@ -21,17 +21,6 @@
                 <show-transaction :tx="tx" :show-tx-info="false"/>
             </div>
 
-            <div v-for="(fee,index) in fees"
-                 :key="`fee_${index}`">
-
-                <div class="row pt-2 pb-2">
-                    <span class="col-5 col-sm-3 fw-medium text-truncate">Fee{{ (fees.length > 1) ? index : ''}}</span>
-                    <span class="col-7 col-sm-9 text-truncate">
-                        <amount :asset="fee.asset" :value="fee.amount" value-class="text-900" :sign="true" />
-                    </span>
-                </div>
-            </div>
-
         </div>
 
     </div>
@@ -51,21 +40,6 @@ export default {
     },
 
     computed:{
-        fees(){
-            if (!this.tx) return []
-
-            if (this.tx.version.eq( PandoraPay.enums.transactions.TransactionVersion.TX_SIMPLE) ){
-                return [ {amount: this.tx.fee, asset: PandoraPay.config.coins.NATIVE_ASSET_FULL_STRING_HEX, feeRate: 1 } ]
-            }else if (this.tx.version.eq( PandoraPay.enums.transactions.TransactionVersion.TX_ZETHER) ) {
-
-                const out = []
-                for (const payload of this.tx.payloads)
-                    out.push({amount: payload.statement.fee.mul(payload.feeRate).div( new Decimal(10).pow( payload.feeLeadingZeros ) ), asset: payload.asset, feeRate: payload.feeRate })
-
-                return out
-            }
-
-        }
     },
